refactor: created and moved features in dedicated folders with a clean architecture
