+++ src/app/admin/[adminKey]/auth/login/page.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
-import { useAdminKeyStore } from "@/lib/stores/storeAdminKey";
+import { useAdminKeyStore } from "@/feature/admin-key/storeAdminKey";
 
 export default function LoginPage({
 	params,
