+++ src/app/api/auth/logout/route.ts
@@ -1,11 +1,1 @@
-import { logout } from "@stralya/auth";
-import { setPrismaClient } from "@stralya/auth";
-import { redirect } from "next/navigation";
-import prisma from "../../../../lib/prisma";
-
-setPrismaClient(prisma);
-
-export async function GET() {
-	await logout();
-	redirect("/");
-}
+export { GET } from "@/feature/user/api/apiUserLogout";
