+++ src/components/providers/UserProvider.tsx
@@ -1,6 +1,6 @@
 "use client";
-import { useCurrentUser } from "@/lib/queries/useUser";
-import { useUserStore } from "@/lib/stores/storeUser";
+import { useCurrentUser } from "@/feature/user/queries/useUser";
+import { useUserStore } from "@/feature/user/stores/storeUser";
 import { useEffect } from "react";
 
 export function UserProvider({ children }: { children: React.ReactNode }) {
