+++ src/feature/user/api/apiUserLogout.ts
@@ -1,0 +1,8 @@
+import { logout } from "@stralya/auth";
+import { redirect } from "next/navigation";
+import "../../../lib/initSetPrismaClient";
+
+export async function GET() {
+	await logout();
+	redirect("/");
+}
