+++ src/lib/initSetPrismaClient.ts
@@ -1,0 +1,4 @@
+import { setPrismaClient } from "@stralya/auth";
+import prisma from "@/lib/prisma";
+
+setPrismaClient(prisma);
